test(research): split PDF furniture region coverage
